Automatically convert swamp platform names to platform names.
Previously these conversions were hard-wired, and the java-cli
would break whenever a new platform became available in a swamp
that wasn't in the fixed mapping table.... even if you didn't
use that platform!

With these changes, the cli is forward-looking with respect to platforms
and any new platform will just work.

There is some work-around if we call rhel "Redhat Enterprise Linux"
instead of rhel .. which is what we do for everything else.   We
do the right thing to convert that spaced name to rhel-ver-bits.